# gradudatednd: optimize code and remove SSE