Rearrange the code in the plain codepath so that the compiler can vectorize it.  By processing groups of four pixels, the compiler is able to vectorize the density computations to gain additional speed.

This iop is fast enough that it becomes memory-bound pretty quickly, but still manages a substantial speedup compared to the existing SSE codepath thanks to the additional vectorization:
```
Thr     SSE     new plain
1      49.61ms 29.14ms   -41.2%
2      27.34ms 14.60ms   -46.6%
4      14.16ms  7.36ms   -48.0%
8       6.89ms  4.11ms   -40.3%
16      5.29ms  4.10ms   -22.4%
32      5.27ms  4.50ms   -14.6%
```
Integration tests show pixel-identical results on tests 0011 and 0012.

I added a local override of `copy_pixel_nontemporal` from darktable.h, since I didn't want to force inclusion of xmmintrin.h in virtually every file by including it from darktable.h.  A better update would be to move it into a different or new header which is only used by iops.  (The function is needed to actually achieve the same or better speed than the SSE code, as using a normal store causes increased memory traffic from cache line loads which the nontemporal store does not generate.)