@TurboGit 2023-02-12
> A better update would be to move it into a different or new header which is only used by iops.

The closet would be `iop_api.h` but it is still included by `imageop.h` and` nlmeans_core.h`